requirements.txt
PyQt5>=5.15.0
PyYAML>=6.0
pydantic>=2.0
psutil>=5.9.0
pyinstaller>=5.0
